requirements.txt
streamlit
numpy
pandas
jsonschema
requests
trimesh
pymongo
python-dotenv
flask
pytest
pytest-cov
pytest-mock
